chore: moved approval hook helpers into relevant directory
